requirements.txt
# Core dependencies
openai==1.70.0
python-dotenv==1.1.0

# Data validation
jsonschema==4.23.0
genson==1.3.0

# Utilities
names_generator==0.2.0
